% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame_difference.R
\name{fd}
\alias{fd}
\title{Frame Differencing}
\usage{
fd(dir, m = 2, cutoff = 1, do.plot = TRUE,
split_screen = TRUE, verbose = FALSE, shuffle = FALSE)
}
\arguments{
\item{dir}{dir specifies a directory containing sequentially numbered images.
Uses current directory by default.}

\item{m}{integer order of the butterworth filter}

\item{cutoff}{real-valued frequency spectrum cutoff proportion for
butterworth filter}

\item{do.plot}{logical indicating whether a plot should be generated}

\item{split_screen}{logical indicating whether frame differencing should
be applied separately to each 1/2 of a screen.}

\item{verbose}{logical indicating whether feedback about which frame is
being anlayzed}

\item{shuffle}{logical indicating whether image order should be shuffled
for the purpose of surrogate analysis}
}
\description{
Performs frame differencing on a sequence of video frames
}
