% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffmpeg.R
\name{ffmpeg}
\alias{ffmpeg}
\title{ffmpeg}
\usage{
ffmpeg(path, filename, fps = NULL, keyframe = FALSE, outfile)
}
\arguments{
\item{path}{character string specifies a path to a video to exported e.g.,
C:/Users/user/video_file_folder/}

\item{filename}{character string of file name including extension e.g.,
'video.mp4'. Most video types are supported.}

\item{fps}{specifies the output frames per second (e.g., 29.97, 8, 1/5)}

\item{keyframe}{logical, if TRUE, fps is ignored and only keyframes are
returned. Hence, sampling rate may not be uniform.}

\item{outfile}{character string giving the output image prefix name}
}
\description{
This function allows you to parse videos into individual frames
}
