% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.spec.R
\name{power.spec}
\alias{power.spec}
\title{Power Spectral Density}
\usage{
power.spec(x, hz = 1, use.period = FALSE,
center = FALSE, detrend = FALSE, pad = TRUE,
logxy = FALSE, do.plot = TRUE, graph.par = NULL)
}
\arguments{
\item{x}{a real valued time series}

\item{hz}{the sampling rate in Hz, default = 1}

\item{use.period}{logical indicating whether power should be given in terms of
period (s) instead of frequency (hz)}

\item{center}{logical indicating whether series should be centered}

\item{detrend}{logical indicating whether linear trend should be removed}

\item{pad}{logical indicating whether zero padding should be applied}

\item{logxy}{should frequency and power be returned in log10 units?}

\item{do.plot}{logical indicating whether plot should be generated. if
logxy = TRUE, plot will also be in log units.}

\item{graph.par}{a list comtaining additional graphical parameters to use for plotting.
The default is NULL}
}
\value{
returns normalized power spectrum
}
\description{
Estimates normalized power spectral density of a time series
}
\examples{
tt = (1:1000)/29.97
sintt = sin(2*pi*tt*3)
ps = power.spec(sintt, hz = 29.97,use.period = FALSE, center = TRUE, detrend = FALSE,
 pad = TRUE, logxy = FALSE, do.plot = TRUE)
}
